refactor: codegen to not be a separated package
