+++ package.json
@@ -6,7 +6,8 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+   "codegen": "graphql-codegen --config codegen.ts"
   },
   "dependencies": {
     "@ethereum-attestation-service/eas-contracts": "^1.7.1",
@@ -30,11 +31,17 @@
     "wagmi": "^2.12.7"
   },
   "devDependencies": {
+    "@graphql-codegen/cli": "5.0.0",
+    "@graphql-codegen/client-preset": "^4.2.6",
+    "@graphql-codegen/typescript": "4.0.1",
+    "@graphql-codegen/typescript-graphql-request": "6.1.0",
+    "@graphql-codegen/typescript-operations": "4.0.1",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "eslint": "^8",
     "eslint-config-next": "14.2.6",
+    "graphql-codegen-plugin-typescript-swr": "0.8.5",
     "pino-pretty": "^11.2.2",
     "postcss": "^8",
     "tailwindcss": "^3.4.1",
