--- package/gql/package.json
@@ -1,31 +1,0 @@
-{
-    "name": "@bacchus-onboarding/gql",
-    "version": "0.0.0",
-    "license": "MIT",
-    "scripts": {
-        "graphql:check-types": "graphql-codegen --config codegen.ts --check",
-        "graphql:update-types": "graphql-codegen --config codegen.ts"
-    },
-    "devDependencies": {
-        "@graphql-codegen/cli": "5.0.0",
-        "@graphql-codegen/client-preset": "^4.2.6",
-        "@graphql-codegen/typescript": "4.0.1",
-        "@graphql-codegen/typescript-graphql-request": "6.1.0",
-        "@graphql-codegen/typescript-operations": "4.0.1",
-        "@types/node": "20.12.7",
-        "@types/react": "18.3.1",
-        "@types/react-dom": "18.3.0",
-        "graphql": "16.8.1",
-        "graphql-codegen-plugin-typescript-swr": "0.8.5",
-        "graphql-request": "6.1.0",
-        "micromatch": "4.0.5",
-        "react": "18.3.1",
-        "react-dom": "18.3.1",
-        "typescript": "5.4.5"
-    },
-    "dependencies": {
-        "graphql-request": "6.1.0",
-        "graphql-tag": "^2.12.6",
-        "swr": "2.2.4"
-    }
-}
